backend/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0

# Database
psycopg2-binary==2.9.7
redis==5.0.1

# API Documentation
flask-restx==1.2.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Background Tasks
celery==5.3.4
flower==2.0.1

# AI/NLP Dependencies
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
openai==1.3.5
sentence-transformers==2.2.2
nltk==3.8.1

# Document Processing
python-docx==0.8.11
PyMuPDF==1.23.8
pdfplumber==0.10.3
python-magic==0.4.27

# PDF Generation
weasyprint==60.1
reportlab==4.0.7

# Payment Processing
stripe==7.8.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
bleach==6.1.0
validators==0.22.0

# Security
bcrypt==4.1.2
cryptography==41.0.8

# Monitoring and Logging
sentry-sdk[flask]==1.38.0

# Development Dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1
